Adding metabo draft example
this introduces an example from metabolomics (WIP) plus a multitude of proposed cv terms corresponding to qc metrics common in clinical analytics (e.g. system suitability tests, but also more common general terms like calibration fit or chromatographic peak asymmetry) used in the example. Minor reencoding of the schema had to be done (content is now an extension of any with the cv attributes copied  in scince inheritance was not workable with a content-type).