Add files via upload
This file represents Dave Tabb's first effort at creating a complete qcML document to the qcML_0_0_10 specification.  This document represents the QuaMeter IDFree metrics for a single mzML file that was peaklisted from a single RAW file titled "1091_Pool_start.raw".  It is marked as version 0.8 to reflect that the individual QuaMeter IDFree metrics have not yet been assigned CV terms.  I also had a question about how to specify the unit for a single-value metric.  This file is not yet validated.