Example qcML via QuaMeter IDFree v0.9
This file represents all the QuaMeter IDFree metrics computed for a single Thermo RAW file, using CV 0.0.10 as a basis.  I have added the PSI-QC-CV terms for all metrics with the following exceptions:
MS1-Count (could not find in this CV revision; this is an obvious candidate for MIAPE)
MS2-Count (could not find in this CV revision; this is an obvious candidate for MIAPE)
MS2-PrecZ-Known (presented as a vector rather than individual metric approach in this CV revision)
MS2-PrecZ-Unknown (presented as a two-part vector rather than individual metric approach in this CV revision)
This file revision also differs from its v0.8 predecessor by giving units for single-value metrics via the unitAccession tag.  I don't understand why we don't label the UO accessions this way with multi-value metrics.